+++ changelog.md
@@ -45,6 +45,7 @@
 - Fixed links to assigned shipping methods were not displayed when editing a rule.
 - Fixed SearchLog did not show top search terms in instant search.
 - Fixed redirect to login page instead of an error message when a guest wants to checkout and anonymous checkout is not allowed.
+- Fixed the category number in category navigation was not displayed when the catalog setting `ShowCategoryProductNumberIncludingSubcategories` was changed.
 
 
 ## Smartstore 5.0.4